done for select from camera
