+++ app/src/main/AndroidManifest.xml
@@ -17,11 +17,11 @@
         android:supportsRtl="true"
         android:networkSecurityConfig="@xml/network_security_config"
         android:theme="@style/AppTheme">
-        <activity android:name=".Main5Activity"></activity>
-        <activity android:name=".Main4Activity" />
-        <activity android:name=".Main3Activity" />
-        <activity android:name=".Main2Activity" />
-        <activity android:name=".MainActivity">
+        <activity android:name="com.example.projekskripsi.Main5Activity"/>
+        <activity android:name="com.example.projekskripsi.SelectFromGalleryActivity" />
+        <activity android:name="com.example.projekskripsi.GetFromCameraActivity" />
+        <activity android:name="com.example.projekskripsi.Main2Activity" />
+        <activity android:name="com.example.projekskripsi.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
