+++ app/src/main/res/layout/activity_main.xml
@@ -16,7 +16,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         android:layout_marginTop="70dp"
-        />
+        android:contentDescription="MainImg" />
 
     <TextView
         android:id="@+id/WelcomeText"
