+++ app/src/main/res/layout/activity_main2.xml
@@ -20,7 +20,8 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="10dp"
             android:adjustViewBounds="true"
-            android:src="@drawable/univ2"></ImageView>
+            android:src="@drawable/univ2"
+            android:contentDescription="TODO"></ImageView>
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -48,7 +49,8 @@
                         android:id="@+id/ambil"
                         android:layout_width="match_parent"
                         android:layout_height="50dp"
-                        android:src="@mipmap/photo"></ImageView>
+                        android:src="@mipmap/photo"
+                        android:contentDescription="TODO"></ImageView>
                     <TextView
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
@@ -80,7 +82,8 @@
                         android:id="@+id/unggah"
                         android:layout_width="match_parent"
                         android:layout_height="50dp"
-                        android:src="@mipmap/browse"></ImageView>
+                        android:src="@mipmap/browse"
+                        android:contentDescription="TODO"></ImageView>
                     
                     <TextView
                         android:layout_width="match_parent"
@@ -126,7 +129,8 @@
                             android:id="@+id/tentang"
                             android:layout_width="match_parent"
                             android:layout_height="50dp"
-                            android:src="@mipmap/about"></ImageView>
+                            android:src="@mipmap/about"
+                            android:contentDescription="TODO"></ImageView>
                         <TextView
                             android:layout_width="match_parent"
                             android:layout_height="wrap_content"
@@ -158,7 +162,8 @@
                             android:id="@+id/keluar"
                             android:layout_width="match_parent"
                             android:layout_height="50dp"
-                           android:src="@mipmap/keluar"></ImageView>
+                           android:src="@mipmap/keluar"
+                            android:contentDescription="TODO"></ImageView>
                         <TextView
                             android:layout_width="match_parent"
                             android:layout_height="wrap_content"
