+++ app/src/main/res/layout/activity_main3.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Main3Activity"
+    tools:context=".GetFromCameraActivity"
     android:paddingBottom="16dp"
     android:paddingLeft="16dp"
     android:paddingRight="16dp"
@@ -16,7 +15,8 @@
         android:id="@+id/image_view3"
         android:scaleType="centerCrop"
         android:layout_width="400dp"
-        android:layout_height="400dp"></ImageView>
+        android:layout_height="400dp"
+        android:contentDescription="TODO" />
     
     <Button
         android:id="@+id/capture_image_btn"
@@ -26,6 +26,6 @@
         android:background="@color/colorPrimary"
         android:textColor="#ffff"
         android:textStyle="bold"
-        android:text="Buka Kamera"></Button>
+        android:text="Buka Kamera"/>
 
 </LinearLayout>
