+++ app/src/main/res/layout/activity_main5.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Main4Activity">
+    tools:context=".SelectFromGalleryActivity">
 
     <de.hdodenhof.circleimageview.CircleImageView
         android:id="@+id/imgview_profile"
@@ -38,7 +38,8 @@
         android:layout_width="match_parent"
         android:layout_height="100dp"
         android:src="@drawable/profileuidesign_2"
-        android:layout_below="@id/desc"></ImageView>
+        android:layout_below="@id/desc"
+        android:contentDescription="TODO"></ImageView>
 
     //Linear layout pertama
 
@@ -54,7 +55,8 @@
             android:layout_width="36dp"
             android:layout_height="36dp"
             android:src="@drawable/ic_person"
-            android:layout_marginLeft="50dp"></ImageView>
+            android:layout_marginLeft="50dp"
+            android:contentDescription="TODO"></ImageView>
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -80,7 +82,8 @@
             android:layout_width="36dp"
             android:layout_height="36dp"
             android:src="@drawable/ic_message"
-            android:layout_marginLeft="50dp"></ImageView>
+            android:layout_marginLeft="50dp"
+            android:contentDescription="TODO"></ImageView>
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -107,7 +110,8 @@
             android:layout_width="36dp"
             android:layout_height="36dp"
             android:src="@drawable/ic_share_black_24dp"
-            android:layout_marginLeft="50dp"></ImageView>
+            android:layout_marginLeft="50dp"
+            android:contentDescription="TODO"></ImageView>
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -133,7 +137,8 @@
             android:layout_width="36dp"
             android:layout_height="36dp"
             android:src="@drawable/ic_location"
-            android:layout_marginLeft="50dp"></ImageView>
+            android:layout_marginLeft="50dp"
+            android:contentDescription="TODO"></ImageView>
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -159,7 +164,8 @@
             android:layout_width="36dp"
             android:layout_height="36dp"
             android:src="@drawable/ic_university"
-            android:layout_marginLeft="50dp"></ImageView>
+            android:layout_marginLeft="50dp"
+            android:contentDescription="TODO"></ImageView>
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
